Move user classes to separate `users` namespace.
