+++ carp.studies.core/src/commonMain/kotlin/dk/cachet/carp/studies/domain/users/Account.kt
@@ -1,4 +1,4 @@
-package dk.cachet.carp.studies.domain
+package dk.cachet.carp.studies.domain.users
 
 import dk.cachet.carp.common.*
