+++ carp.studies.core/src/commonTest/kotlin/dk/cachet/carp/studies/domain/users/InMemoryUserRepository.kt
@@ -1,4 +1,4 @@
-package dk.cachet.carp.studies.domain
+package dk.cachet.carp.studies.domain.users
 
 import dk.cachet.carp.common.*
